change int to long for SSN Sofia + Alexandra

+++ src/com/Account.java
@@ -9,7 +9,7 @@ public class Account {
 
     private LinkedHashMap<String, String> details = new LinkedHashMap<>();
 
-    public Account(int accountNumber, double accountBalance, double debt, int ssn) {
+    public Account(int accountNumber, double accountBalance, double debt, long ssn) {
         this.details.put("accountnumber", String.valueOf(accountNumber));
         this.details.put("accountbalance", String.valueOf(accountBalance));
         this.details.put("debt", String.valueOf(debt));
@@ -40,12 +40,12 @@ public class Account {
         return Double.parseDouble(details.get("debt"));
     }
 
-    public void setSsn(int ssn) {
+    public void setSsn(long ssn) {
         details.put("ssn", String.valueOf(ssn));
     }
 
-    public int getSsn() {
-        return Integer.parseInt(details.get("ssn"));
+    public long getSsn() {
+        return Long.parseLong(details.get("ssn"));
     }
 
     public List<String> getList() {
